pyproject.toml
[project]
name = "grocery-agent"
version = "0.1.0"
description = "Pittsburgh grocery price comparison and route planning AI agent"
requires-python = ">=3.11"

dependencies = [
    "google-genai>=1.0.0",   # Gemma 3 via Google GenAI SDK
    "requests>=2.31.0",              # ORS API calls and scraping
    "beautifulsoup4>=4.12.0",        # HTML parsing for price scraper
    "playwright>=1.44.0",            # Headless browser for JS-heavy store sites
    "pytest>=8.0.0",                 # Test runner
    "pytest-mock>=3.14.0",          # Mocking in tests
    "python-dotenv>=1.0.0",          # Load .env for API keys
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]                   # Make project root importable in tests

[tool.hatch.build.targets.wheel]
packages = ["agent", "tools", "config"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"